--- shared/README.md
@@ -1,3 +1,0 @@
-# SHARED
-
-This is not a standard directory. This is for common functionality that can be reused across pages, components, and server functions.
