+++ shared/bible/core/display.ts
