+++ shared/bible/core/encoding.ts
