+++ shared/bible/core/metadata.ts
