+++ shared/bible/core/ranges.ts
