+++ shared/bible/core/segments.ts
