+++ shared/bible/core/validation.ts
