+++ shared/bible/scripture.ts
