+++ shared/bible/static/bible-books.ts
