+++ shared/bible/static/chapter-verses/nasb.ts
