+++ shared/insights/calendar.ts
