+++ shared/insights/charts.test.ts
