+++ shared/insights/charts.ts
