+++ shared/insights/insights.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from 'vitest';
 import dayjs from 'dayjs';
-import Bible from './bible';
+import Bible from '../bible';
 import {
   buildContributionCalendar,
   computeBookFrequencies,
