+++ shared/insights/reading-suggestions.ts
@@ -1,5 +1,5 @@
 import dayjs from 'dayjs';
-import Bible from './bible';
+import Bible from '../bible';
 
 /**
  * Reading-suggestion algorithm shared by the web and mobile apps.
