+++ shared/platform/browser-cache.ts
