+++ shared/platform/device.ts
