+++ shared/platform/translator.ts
